zec_api/services/user-service/requirements.txt
fastapi[standard]
sqlalchemy
pydantic
pydantic-settings
typing-extensions
uvicorn[standard]
psycopg[binary]
requests
PyJWT